+++ modules/loader-utils/package.json
@@ -28,18 +28,14 @@
     "README.md"
   ],
   "browser": {
-    "./src/node/to-array-buffer.node.js": false,
-    "./dist/es5/node/to-array-buffer.node.js": false,
-    "./dist/esm/node/to-array-buffer.node.js": false,
-    "./dist/es6/node/to-array-buffer.node.js": false,
-    "./src/node/to-buffer.node.js": false,
-    "./dist/es5/node/to-buffer.node.js": false,
-    "./dist/esm/node/to-buffer.node.js": false,
-    "./dist/es6/node/to-buffer.node.js": false,
-    "./src/lib/library-utils/require-utils.node.js": false,
-    "./dist/es5/lib/library-utils/require-utils.node.js": false,
-    "./dist/esm/lib/library-utils/require-utils.node.js": false,
-    "./dist/es6/lib/library-utils/require-utils.node.js": false
+    "./src/lib/node/buffer-utils.node.js": false,
+    "./dist/es5/lib/node/buffer-utils.node.js": false,
+    "./dist/esm/lib/node/buffer-utils.node.js": false,
+    "./dist/es6/lib/node/buffer-utils.node.js": false,
+    "./src/lib/node/library-utils/require-utils.node.js": false,
+    "./dist/es5/lib/node/library-utils/require-utils.node.js": false,
+    "./dist/esm/lib/node/library-utils/require-utils.node.js": false,
+    "./dist/es6/lib/node/library-utils/require-utils.node.js": false
   },
   "scripts": {
     "pre-build-disabled": "npm run build-bundle && npm run build-bundle -- --env.dev",
